fixed socket.io for leaving a room

+++ server/server-socket.js
@@ -283,35 +283,18 @@ module.exports = {
           await cleanupRooms();
 
           const roomCode = generateRoomCode();
-          console.log("Generated new room code:", roomCode);
-
           const newRoom = new Room({
             code: roomCode,
             hostId: socket.id,
             players: [{ id: socket.id, name: playerName }],
           });
-
           await newRoom.save();
-          console.log("Room saved successfully:", {
-            code: newRoom.code,
-            hostId: newRoom.hostId,
-            players: newRoom.players,
-            _id: newRoom._id,
-          });
 
-          // Verify room was saved
-          const verifyRoom = await Room.findOne({ code: roomCode });
-          console.log(
-            "Verification - Found room in DB:",
-            verifyRoom
-              ? {
-                  code: verifyRoom.code,
-                  hostId: verifyRoom.hostId,
-                  players: verifyRoom.players,
-                  _id: verifyRoom._id,
-                }
-              : "Not found"
-          );
+          // In memory game state
+          rooms[roomCode] = {
+            players: { [socket.id]: playerName },
+            scores: { [socket.id]: 0 },
+          };
 
           socket.join(roomCode);
 
@@ -335,6 +318,7 @@ module.exports = {
             return callback({ error: "Room code and player name are required." });
           }
 
+          // Save in MONGODB
           const room = await Room.findOne({ code: roomCode });
           if (!room) {
             return callback({ error: "Room not found." });
@@ -342,15 +326,6 @@ module.exports = {
           if (room.isGameStarted) {
             return callback({ error: "Game has already started in this room." });
           }
-
-          console.log(`\n[Room ${roomCode}] Player attempting to join:`, {
-            name: playerName,
-            socketId: socket.id,
-          });
-          console.log(
-            `Current players in room:`,
-            room.players.map((p) => ({ name: p.name, id: p.id }))
-          );
 
           // Check if player is already in the room
           const existingPlayer = room.players.find((player) => player.name === playerName);
@@ -366,28 +341,17 @@ module.exports = {
             // Add the player to room in database only if they're not already there
             room.players.push({ id: socket.id, name: playerName });
           }
+          await room.save();
 
-          // Add player to the in-memory room data
+          // Add player to the IN-MEMORY room data
           if (!rooms[roomCode]) {
             rooms[roomCode] = { players: {}, scores: {} };
           }
           rooms[roomCode].players[socket.id] = playerName;
+          rooms[roomCode].scores[socket.id] = 0;
 
-          if (!rooms[roomCode].scores[socket.id]) {
-            rooms[roomCode].scores[socket.id] = 0;
-          }
-
-          await room.save();
+          // SOCKETIO
           socket.join(roomCode);
-
-          console.log(
-            `\n[Room ${roomCode}] Updated player list:`,
-            room.players.map((p) => ({
-              name: p.name,
-              id: p.id,
-              isHost: p.id === room.hostId,
-            }))
-          );
 
           // Immediately emit room data to all clients in the room
           io.to(roomCode).emit("roomData", {
@@ -410,38 +374,44 @@ module.exports = {
         }
 
         try {
+          // MONGODB cleanup
           const room = await Room.findOne({ code: roomCode });
           if (!room) {
             return callback({ error: "Room not found." });
           }
 
-          console.log(`\n[Room ${roomCode}] Player leaving:`, { socketId: socket.id });
-          console.log(
-            `Players before removal:`,
-            room.players.map((p) => ({ name: p.name, id: p.id }))
-          );
-
           room.players = room.players.filter((player) => player.id !== socket.id);
-
           if (room.hostId === socket.id) {
-            // Options are to either delete the room or assign a new host (if there are other players left)
-            await Room.deleteOne({ code: roomCode });
-            socket.leave(roomCode);
-            return callback({ success: true });
+            if (room.players.length > 0) {
+              // other players exist in room- assign new host
+              room.hostId = room.players[0].id;
+              await room.save();
+            } else {
+              // Host was last player - delete room
+              await Room.deleteOne({ code: roomCode });
+            }
+          } else {
+            // not host- just save player removal
+            await room.save();
           }
 
-          await room.save();
+          // In Memory game state cleanup
+          if (rooms[roomCode]) {
+            delete rooms[roomCode].players[socket.id];
+            if (rooms[roomCode].hostId === socket.id) {
+              const remainingPlayers = Object.keys(rooms[roomCode].players);
+              if (remainingPlayers.length > 0) {
+                rooms[roomCode].hostId = remainingPlayers[0];
+              } else {
+                delete rooms[roomCode];
+              }
+            }
+          }
+
+          // SOCKETIO cleanup
           socket.leave(roomCode);
 
-          console.log(
-            `\n[Room ${roomCode}] Updated player list:`,
-            room.players.map((p) => ({
-              name: p.name,
-              id: p.id,
-              isHost: p.id === room.hostId,
-            }))
-          );
-
+          // Notify all users in room of the leaving user
           io.to(roomCode).emit("roomData", {
             players: room.players,
             hostId: room.hostId,
@@ -496,14 +466,14 @@ module.exports = {
       //   const rooms = await Room.find({ "players.id": socket.id });
       //   for (const room of rooms) { // TODO FIX THIS SO DON'T HAVE MANY OLD ROOMS
       //     if (room.hostId === socket.id) {
-      //       // Options are to either delete the room or assign a new host (if there are other players left) 
+      //       // Options are to either delete the room or assign a new host (if there are other players left)
       //       await Room.deleteOne({ code: room.code });
       //       socket.leave(room.code);
       //       return;
       //     }
       //     room.players = room.players.filter(player => player.id !== socket.id);
       //     await room.save();
-      //     socket.leave(room.code);  
+      //     socket.leave(room.code);
       //     io.to(room.code).emit("roomData", {
       //       players: room.players,
       //       hostId: room.hostId,
